// packages/angular/src/app/shared/api/model/login-request.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Web System Samples
 * Webシステム開発の学習用サンプルアプリケーションです。
 * OpenAPI spec version: 0.0.0
 */

/**
 * ログインAPIのリクエストボディ
 */
export interface LoginRequest {
    /**
     * アカウント名
     * @minLength 1
     */
    account: string;
    /**
     * パスワード
     * @minLength 1
     */
    password: string;
}